pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: Apache Software License",
]
name = "nomad-ait-echt-oasis"
description = "A NOMAD plugin containing schemas, parsers, etc. for material processing and analysis methods at the ECHT group of AIT."
version = "0.2.1"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Maximilian Wolf", email = "maximilian.wolf@ait.ac.at" },
]
maintainers = [
    { name = "Maximilian Wolf", email = "maximilian.wolf@ait.ac.at" },
]
license = { file = "LICENSE" }
dependencies = [
    "nomad-lab>=1.3.0",
    "python-magic-bin; sys_platform == 'win32'",
    "fastapi>=0.118.0",
    "nomad-material-processing @ git+https://github.com/FAIRmat-NFDI/nomad-material-processing.git"
]

[project.urls]
Repository = "https://github.com/mxwalbert/ait-echt-oasis"

[project.optional-dependencies]
dev = ["ruff", "pytest", "structlog"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint
    "PL",
]

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
    "PLC0415", # `import` should be at the top-level of a file
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points.'nomad.plugin']
schema_package_infrastructure = "nomad_ait_echt_oasis.schema_packages:infrastructure"
schema_package_vapor_deposition = "nomad_ait_echt_oasis.schema_packages:vapor_deposition"
schema_package_physical_vapor_deposition = "nomad_ait_echt_oasis.schema_packages:physical_vapor_deposition"
schema_package_sputter_deposition = "nomad_ait_echt_oasis.schema_packages:sputter_deposition"
parser_entry_point = "nomad_ait_echt_oasis.parsers:parser_entry_point"
normalizer_entry_point = "nomad_ait_echt_oasis.normalizers:normalizer_entry_point"
app_entry_point = "nomad_ait_echt_oasis.apps:app_entry_point"
example_upload_entry_point = "nomad_ait_echt_oasis.example_uploads:example_upload_entry_point"
api_entry_point = "nomad_ait_echt_oasis.apis:api_entry_point"

[tool.cruft]
# Avoid updating workflow files, this leads to permissions issues
skip = [".github/*"]
